excalidraw-app/backend/requirements.txt
# FastAPI backend dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
